feat: 增加Map清空与弹出同步操作

+++ kbe/src/lib/entitydef/fixeddict.cpp
@@ -281,7 +281,7 @@ void FixedDict::onInstallScript(PyObject* mod)
 }
 
 //-------------------------------------------------------------------------------------
-int FixedDict::mp_length(PyObject* self)
+Py_ssize_t FixedDict::mp_length(PyObject* self)
 {
 	return PyDict_Size(static_cast<FixedDict*>(self)->pyDict_);
 }
