+++ kbe/src/lib/entitydef/fixeddict.h
@@ -68,7 +68,7 @@ public:
 	static int mp_ass_subscript(PyObject* self, PyObject* key, 
 		PyObject* value);
 
-	static int mp_length(PyObject* self);
+	static Py_ssize_t mp_length(PyObject* self);
 
 	/** 
 		初始化固定字典
